+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 0.1.29.dev0 (Work In Progress)
 
 - add CUDA_MODULE_LOADING=EAGER to force CUDA kernel preloading
+- drop the no-longer working pytest-ipynb, and switch to nbmake for testing
 
 
 ## 0.1.28 (2021-12-07)